Distributed conditional logic for state chaning to state classes
Rather than have the Player controller broadly handle all input and state changing, I learned that it's better to have each state handle which states it can enter or exit. This separation of concern allows for easier implementation of the different states and allows the passing of the same Player controller object between states. However, there is still a concern about how the GroundCheck class updates the player and how Unity's component based architecture might influence updating in the future.