Fixed camera movement bug
Previously the camera's z-index and FoV were based on an arbitrary horizontal distance which was calculated using a partial distance formula. This broke the FoV when 2 or more players are stacked on top of each other (and have a max distance apart of close to 0). The solution was to base the z-index on a combined max distance apart using the full distance formula.