Better encapsulated state logic and update Unity
All behavior for particular states now resides directly in the state that it corresponds too. However, this led to a bug in which the JumpingState was reverting back to the previous state before the groundcheck logic had finished. This was resolved with C#'s async delay functionality (possibly only a temporary fix). Jump forces have been replaced with velocities for more responsive movement

Additionally, Unity and subsequently C# have been updated to the latest versions.