Rewrote Combo System and Modified Dependencies
The previous 'ComboManager' was buggy and overimplemented. After much research, I discovered a simpler solution online that can take any number of combos and is tied to the combo object. Related dependencies were tweaked and/or deleted and given more logical/appropriate names. Additionally, a new Slam state was created.