Made JoyCombo and Combo Class More Robust and Added Two New States
A Combo class is now comprised of a Double Tuple Array opposed to the simple string array. The purpose is to specify a Type to the input (button, leftjoy, trigger,etc.) JoyCombo (formally GameControllerCombo) now does a broader check before checking for particulars. This simplifies the single huge if-statement and makes for a cleaner check.

Two new states, pebble shot and crystal shard, have been created but are not fleshed out. Additionally, minor cleanup of the project has been done and the README has been updated.